+++ tests/src/sanity.rs
@@ -5,9 +5,10 @@ mod tests {
     use std::{
         collections::HashMap,
         fmt::Debug,
-        net::{Ipv4Addr, UdpSocket},
+        net::Ipv4Addr,
         process::Stdio,
         sync::{Arc, Mutex},
+        thread::sleep,
         time::Duration,
     };
 
@@ -19,7 +20,7 @@ mod tests {
         core::v1::{Pod, Service},
     };
     use kube::{
-        api::{DeleteParams, ListParams, PostParams},
+        api::{DeleteParams, ListParams, LogParams, PostParams},
         core::WatchEvent,
         runtime::wait::{await_condition, conditions::is_pod_running},
         Api, Client, Config,
@@ -40,6 +41,7 @@ mod tests {
     use tokio_util::sync::{CancellationToken, DropGuard};
 
     static TEXT: &str = "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.";
+    const CONTAINER_NAME: &str = "test";
 
     pub async fn watch_resource_exists<K: Debug + Clone + DeserializeOwned>(
         api: &Api<K>,
@@ -286,7 +288,7 @@ mod tests {
         env.insert("MIRRORD_AGENT_IMAGE", "test");
         env.insert("MIRRORD_CHECK_VERSION", "false");
         env.insert("MIRRORD_AGENT_RUST_LOG", "warn,mirrord=debug");
-        env.insert("MIRRORD_IMPERSONATED_CONTAINER_NAME", "test");
+        env.insert("MIRRORD_IMPERSONATED_CONTAINER_NAME", CONTAINER_NAME);
         env.insert("MIRRORD_AGENT_COMMUNICATION_TIMEOUT", "180");
         env.insert("RUST_LOG", "warn,mirrord=debug");
         let server = Command::new(path)
@@ -374,6 +376,8 @@ mod tests {
     async fn service(
         #[future] kube_client: Client,
         #[default("default")] namespace: &str,
+        #[default("NodePort")] service_type: &str,
+        #[default("ghcr.io/metalbear-co/mirrord-pytest:latest")] image: &str,
     ) -> EchoService {
         let kube_client = kube_client.await;
         let deployment_api: Api<Deployment> = Api::namespaced(kube_client.clone(), namespace);
@@ -383,29 +387,29 @@ mod tests {
             "apiVersion": "apps/v1",
             "kind": "Deployment",
             "metadata": {
-                "name": name,
+                "name": &name,
                 "labels": {
-                    "app": name
+                    "app": &name
                 }
             },
             "spec": {
                 "replicas": 1,
                 "selector": {
                     "matchLabels": {
-                        "app": name
+                        "app": &name
                     }
                 },
                 "template": {
                     "metadata": {
                         "labels": {
-                            "app": name
+                            "app": &name
                         }
                     },
                     "spec": {
                         "containers": [
                             {
-                                "name": "test",
-                                "image": "ghcr.io/metalbear-co/mirrord-pytest:latest",
+                                "name": &CONTAINER_NAME,
+                                "image": &image,
                                 "ports": [
                                     {
                                         "containerPort": 80
@@ -424,7 +428,7 @@ mod tests {
                                         "name": "MIRRORD_FAKE_VAR_THIRD",
                                         "value": "foo=bar"
                                     }
-                                  ]
+                                ],
                             }
                         ]
                     }
@@ -446,17 +450,23 @@ mod tests {
                 }
             },
             "spec": {
-                "type": "NodePort",
+                "type": &service_type,
                 "selector": {
                     "app": &name
                 },
                 "sessionAffinity": "None",
                 "ports": [
                     {
+                        "name": "udp",
+                        "protocol": "UDP",
+                        "port": 31415,
+                    },
+                    {
+                        "name": "http",
                         "protocol": "TCP",
                         "port": 80,
-                        "targetPort": 80
-                    }
+                        "targetPort": 80,
+                    },
                 ]
             }
         }))
@@ -480,6 +490,20 @@ mod tests {
             _pod: pod_guard,
             _service: service_guard,
         }
+    }
+
+    /// Service that should only be reachable from inside the cluster, as a communication partner
+    /// for testing outgoing traffic. If this service receives the application's messages, they
+    /// must have been intercepted and forwarded via the agent to be sent from the impersonated pod.
+    #[fixture]
+    async fn udp_logger_service(#[future] kube_client: Client) -> EchoService {
+        service(
+            kube_client,
+            "default",
+            "ClusterIP",
+            "ghcr.io/metalbear-co/mirrord-node-udp-logger:latest",
+        )
+        .await
     }
 
     fn resolve_node_host() -> String {
@@ -970,11 +994,9 @@ mod tests {
         process.assert_stderr();
     }
 
-    // TODO: This is valid for all "outgoing" tests:
-    // We have no way of knowing if they're actually being hooked, so they'll pass without mirrord,
-    // which is bad.
-    // An idea to solve this problem would be to have some internal (or test-only) specific messages
-    // that we can pass back and forth between `layer` and `agent`.
+    // TODO: change outgoing TCP tests to use the same setup as in the outgoing UDP test so that
+    //       they actually verify that the traffic is intercepted and forwarded (and isn't just
+    //       directly sent out from the local application).
     #[rstest]
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
     pub async fn test_outgoing_traffic_single_request_enabled(#[future] service: EchoService) {
@@ -1067,30 +1089,80 @@ mod tests {
         process.assert_stderr();
     }
 
+    /// Currently, mirrord only intercepts and forwards outgoing udp traffic if the application
+    /// binds a non-0 port and calls `connect`. This test runs with mirrord a node app that does
+    /// that and verifies that mirrord intercepts and forwards the outgoing udp message.
     #[rstest]
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    #[timeout(Duration::from_secs(30))]
-    pub async fn test_outgoing_traffic_udp(#[future] service: EchoService) {
-        let service = service.await;
-        let socket = UdpSocket::bind("127.0.0.1:0").unwrap();
-        let port = socket.local_addr().unwrap().port().to_string();
+    #[timeout(Duration::from_secs(240))]
+    pub async fn test_outgoing_traffic_udp_with_connect(
+        #[future] udp_logger_service: EchoService,
+        #[future] service: EchoService,
+        #[future] kube_client: Client,
+    ) {
+        let internal_service = udp_logger_service.await; // Only reachable from withing the cluster.
+        let target_service = service.await; // Impersonate a pod of this service, to reach internal.
+        let kube_client = kube_client.await;
+        let pod_api: Api<Pod> = Api::namespaced(kube_client.clone(), &internal_service.namespace);
+        let mut lp = LogParams {
+            container: Some(String::from(CONTAINER_NAME)),
+            follow: false,
+            limit_bytes: None,
+            pretty: false,
+            previous: false,
+            since_seconds: None,
+            tail_lines: None,
+            timestamps: false,
+        };
 
         let node_command = vec![
             "node",
-            "node-e2e/outgoing/test_outgoing_traffic_udp_client.mjs",
-            &port,
+            "node-e2e/outgoing/test_outgoing_traffic_udp_client_with_connect.mjs",
+            "31415",
+            // Reaching service by only service name is only possible from within the cluster.
+            &internal_service.name,
         ];
-        let mut process = run(node_command, &service.pod_name, None, None).await;
 
-        // Listen for UDP message from mirrorded app.
-        let mut buf = [0; 27];
-        let amt = socket.recv(&mut buf).unwrap();
-        assert_eq!(amt, 27);
-        assert_eq!(buf, "Can I pass the test please?".as_ref()); // Sure you can.
+        // Meta-test: verify that the application cannot reach the internal service without
+        // mirrord forwarding outgoing UDP traffic via the target pod.
+        // If this verification fails, the test itself is invalid.
+        let mirrord_no_outgoing = vec!["--no-outgoing"];
+        let mut process = run(
+            node_command.clone(),
+            &target_service.pod_name,
+            Some(&target_service.namespace),
+            Some(mirrord_no_outgoing),
+        )
+        .await;
+        let res = process.child.wait().await.unwrap();
+        assert!(!res.success()); // Should fail because local process cannot reach service.
+        let logs = pod_api.logs(&internal_service.pod_name, &lp).await;
+        assert_eq!(logs.unwrap(), "");
 
+        // Run mirrord with outgoing enabled.
+        let mut process = run(
+            node_command,
+            &target_service.pod_name,
+            Some(&target_service.namespace),
+            None,
+        )
+        .await;
         let res = process.child.wait().await.unwrap();
         assert!(res.success());
         process.assert_stderr();
+
+        // Verify that the UDP message sent by the application reached the internal service.
+        lp.follow = true; // Follow log stream.
+        let logs = pod_api
+            .log_stream(&internal_service.pod_name, &lp)
+            .await
+            .unwrap()
+            .try_next()
+            .await
+            .unwrap()
+            .unwrap();
+        let logs = String::from_utf8_lossy(&logs);
+        assert!(logs.contains("Can I pass the test please?")); // Of course you can.
     }
 
     #[rstest]